--- src/App.scss
@@ -1,76 +1,0 @@
-/* #root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
-}
-
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-}
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-.logo.react:hover {
-  filter: drop-shadow(0 0 2em #61dafbaa);
-}
-
-@keyframes logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  a:nth-of-type(2) .logo {
-    animation: logo-spin infinite 20s linear;
-  }
-}
-
-.card {
-  padding: 2em;
-}
-
-.read-the-docs {
-  color: #888;
-} */
-
-// Colors
-$bg-color: var(--chakra-colors-gray-900);
-$dot-color: var(--chakra-colors-gray-600);
-
-// Dimensions
-$dot-size: 2px;
-$dot-space: 22px;
-
-body {
-    background: $bg-color;
-    // min-height: 100vh;
-    // display: flex;
-    // align-items: center;
-    // justify-content: center;
-    // font-family: 'Roboto', sans-serif;
-    // font-size: 1.5rem;
-    // color: #fff;
-}
-
-.background {
-    background: linear-gradient(90deg, $bg-color ($dot-space - $dot-size), transparent 1%) center,
-        linear-gradient($bg-color ($dot-space - $dot-size), transparent 1%) center, $dot-color;
-    background-size: $dot-space $dot-space;
-    animation: moveDots 4s linear infinite;
-}
-
-@keyframes moveDots {
-    from {
-        background-position: 0 0;
-    }
-    to {
-        background-position: $dot-space $dot-space;
-    }
-}
