--- src/Link.tsx
@@ -1,16 +1,0 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
-import React from 'react';
-import { usePageContext } from './usePageContext';
-
-function Link(props: { href?: string; className?: string; children: React.ReactNode }) {
-    const pageContext = usePageContext();
-    const className = [props.className, pageContext.urlPathname === props.href && 'is-active'].filter(Boolean).join(' ');
-    return (
-        <a
-            {...props}
-            className={className}
-        />
-    );
-}
-
-export { Link };
