+++ src/components/PageShell/PageShell.scss
@@ -1,23 +1,3 @@
-/* This CSS is common to all pages */
-
-// body {
-//     margin: 0;
-//     font-family: sans-serif;
-// }
-// * {
-//     box-sizing: border-box;
-// }
-// a {
-//     text-decoration: none;
-// }
-
-// .navitem {
-//     padding: 3px 10px;
-// }
-// .navitem.is-active {
-//     background-color: #eee;
-// }
-
 // Colors
 $bg-color: var(--chakra-colors-gray-900);
 $dot-color: var(--chakra-colors-gray-600);
