+++ src/components/buttons/DiscordButton.tsx
@@ -2,7 +2,7 @@
 import { Button, Link } from '@chakra-ui/react';
 import { memo } from 'react';
 import { SiDiscord } from 'react-icons/si/index.js';
-import { discordLink } from '../utils/links';
+import { discordLink } from '../../utils/links';
 
 export const DiscordButton = memo(() => {
     return (
