+++ src/components/buttons/GitHubButton.tsx
@@ -2,7 +2,7 @@
 import { Button, Link, HStack, Text, Tag, Icon } from '@chakra-ui/react';
 import { memo } from 'react';
 import { BsFillStarFill, BsGithub } from 'react-icons/bs/index.js';
-import { githubLink } from '../utils/links';
+import { githubLink } from '../../utils/links';
 
 export const GitHubButton = memo(({ stars }: { stars?: number }) => {
     return (
