+++ src/pages/download.page.tsx
@@ -1,20 +1,18 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
-import { useEffect, useState } from 'react';
 import banner from '../assets/banner.png';
-import '../App.scss';
-import { Box, Button, Flex, HStack, Icon, Image, Link, Spacer, Text, VStack } from '@chakra-ui/react';
-import { getRepoInfo } from '../api/api';
+import '../index.scss';
+import { Box, Button, HStack, Icon, Image, Link, Spacer, Text, VStack } from '@chakra-ui/react';
 import { IGithubRelease, IReleaseAsset } from '../types/githubTypes';
-import { isSupportedOS, OS } from '../utils';
+import { OS } from '../utils';
 import { BsWindows, BsApple, BsFillQuestionDiamondFill } from 'react-icons/bs/index.js';
 import { FaLinux } from 'react-icons/fa/index.js';
 import { MdDownload } from 'react-icons/md/index.js';
 import ReactMarkdown from 'react-markdown';
 import ChakraUIRenderer from 'chakra-ui-markdown-renderer';
-import { GitHubButton } from '../components/GitHubButton';
-import { DiscordButton } from '../components/DiscordButton';
-import { KofiButton } from '../components/KofiButton';
-import { ShellWrapper } from '../PageShell';
+import { GitHubButton } from '../components/buttons/GitHubButton';
+import { DiscordButton } from '../components/buttons/DiscordButton';
+import { KofiButton } from '../components/buttons/KofiButton';
+import { ShellWrapper } from '../components/PageShell/PageShell';
 
 function Page(pageProps: {
     latestVersion: IGithubRelease;
@@ -65,85 +63,86 @@ function Page(pageProps: {
 
     return (
         <ShellWrapper>
-            <div>
-                <VStack spacing={8}>
-                    <HStack
-                        spacing={6}
-                        w="full"
+            <VStack
+                spacing={8}
+                mb="auto"
+            >
+                <HStack
+                    spacing={6}
+                    w="full"
+                >
+                    <Image
+                        src={banner}
+                        w="360px"
+                    />
+                    <Spacer></Spacer>
+                    <GitHubButton />
+                    <DiscordButton />
+                    <KofiButton />
+                </HStack>
+                <VStack>
+                    <Button
+                        colorScheme="green"
+                        w={512}
+                        h={32}
+                        borderRadius="2xl"
+                        onClick={() => {
+                            if (currentBuild != null) {
+                                window.location.href = currentBuild?.browser_download_url;
+                            }
+                        }}
                     >
-                        <Image
-                            src={banner}
-                            w="360px"
-                        />
-                        <Spacer></Spacer>
-                        <GitHubButton />
-                        <DiscordButton />
-                        <KofiButton />
-                    </HStack>
-                    <VStack>
-                        <Button
-                            colorScheme="green"
-                            w={512}
-                            h={32}
-                            borderRadius="2xl"
-                            onClick={() => {
-                                if (currentBuild != null) {
-                                    window.location.href = currentBuild?.browser_download_url;
-                                }
-                            }}
+                        <VStack>
+                            <HStack>
+                                <Icon
+                                    boxSize={8}
+                                    as={MdDownload}
+                                ></Icon>
+                                <Text
+                                    fontSize={36}
+                                    fontWeight="bold"
+                                >
+                                    {latestVersion != null ? `Download ${latestVersion?.name}` : 'Loading...'}
+                                </Text>
+                            </HStack>
+                            <HStack>
+                                <Icon as={icon}></Icon>
+                                <Text fontSize={18}>{OS.name}</Text>
+                            </HStack>
+                        </VStack>
+                    </Button>
+                    <Text color="white">
+                        Or download the{' '}
+                        <Link
+                            color="blue.300"
+                            href={zipBuild?.browser_download_url}
                         >
-                            <VStack>
-                                <HStack>
-                                    <Icon
-                                        boxSize={8}
-                                        as={MdDownload}
-                                    ></Icon>
-                                    <Text
-                                        fontSize={36}
-                                        fontWeight="bold"
-                                    >
-                                        {latestVersion != null ? `Download ${latestVersion?.name}` : 'Loading...'}
-                                    </Text>
-                                </HStack>
-                                <HStack>
-                                    <Icon as={icon}></Icon>
-                                    <Text fontSize={18}>{OS.name}</Text>
-                                </HStack>
-                            </VStack>
-                        </Button>
-                        <Text color="white">
-                            Or download the{' '}
-                            <Link
-                                color="blue.300"
-                                href={zipBuild?.browser_download_url}
-                            >
-                                portable version (zip)
-                            </Link>
-                        </Text>
-                    </VStack>
-                    <Box
-                        color="white"
-                        w="full"
-                        bgColor="gray.700"
-                        borderRadius="lg"
-                    >
-                        <Box
-                            mx={2}
-                            p={4}
-                            h="49rem"
-                            w="full"
-                            overflowY="scroll"
-                        >
-                            <ReactMarkdown
-                                components={ChakraUIRenderer()}
-                                skipHtml
-                            >
-                                {computedChangelog}
-                            </ReactMarkdown>
-                        </Box>
-                    </Box>
+                            portable version (zip)
+                        </Link>
+                    </Text>
                 </VStack>
-            </div>
+                <Box
+                    color="white"
+                    w="full"
+                    bgColor="gray.700"
+                    borderRadius="lg"
+                >
+                    <Box
+                        mx={2}
+                        p={4}
+                        h="49rem"
+                        w="full"
+                        overflowY="scroll"
+                    >
+                        <ReactMarkdown
+                            components={ChakraUIRenderer()}
+                            skipHtml
+                        >
+                            {computedChangelog}
+                        </ReactMarkdown>
+                    </Box>
+                </Box>
+            </VStack>
         </ShellWrapper>
     );
 }
